Move symmetric and equivariant layers to their own file
